src/django/requirements.txt
djangorestframework==3.5.3
django-localflavor==1.4.1
django-storages==1.5.2
celery[sqs]==4.0
boto==2.43.0
us==0.9.1
